refactor: rewrite playground in ts
