+++ index.html
@@ -130,7 +130,7 @@
     </svg>
   </a>
   <script type="module" src="./web-component/index.js"></script>
-  <script type="module" src="./playground/playground.js"></script>
+  <script type="module" src="./playground/main.ts"></script>
 </body>
 
 </html>
